Added transaction reference count. Cleaned up API.

+++ MamothDB.Server/Core/Models/MetaTransaction.cs
@@ -12,6 +12,7 @@ namespace MamothDB.Server.Core.Models
         private ServerCore _core;
         private MetaLatchKeyCollection _latchKeys = new MetaLatchKeyCollection();
         private HashSet<string> transactedItems = new HashSet<string>();
+        private int _references = 0;
 
         /// <summary>
         /// Implicit transactions were created by the engine and must be completed at the end of any action.
@@ -55,6 +56,10 @@ namespace MamothDB.Server.Core.Models
             }
         }
 
+        /// <summary>
+        /// Adds a key to the transaction for an outstanding latch.
+        /// </summary>
+        /// <param name="latchKey"></param>
         public void AddLatchKey(MetaLatchKey latchKey)
         {
             _latchKeys.Add(latchKey);
@@ -102,21 +107,44 @@ namespace MamothDB.Server.Core.Models
             });
         }
 
-        public void Commit()
+        /// <summary>
+        /// Increases the referecne count of the current transaction.
+        /// </summary>
+        public void Enlist()
         {
-            _latchKeys.TurnInAllKeys();
-
-            if (_session != null)
-            {
-                _session.CurrentTransaction = null;
-            }
-
-            if (Directory.Exists(TransactionBackupPath))
-            {
-                Directory.Delete(TransactionBackupPath, true);
-            }
+            _references++;
         }
 
+        /// <summary>
+        /// Decreases the transaction reference count. If it falls to zero, the transaction is comitted.
+        /// </summary>
+        /// <returns></returns>
+        public bool Commit()
+        {
+            _references--;
+
+            if (_references < 0)
+            {
+                throw new Exception("Transaction reference count fell below zero.");
+            }
+
+            if (_references == 0)
+            {
+                _latchKeys.TurnInAllKeys();
+
+                if (Directory.Exists(TransactionBackupPath))
+                {
+                    Directory.Delete(TransactionBackupPath, true);
+                }
+                return true;
+            }
+
+            return false;
+        }
+
+        /// <summary>
+        /// Unconditionally rolls back the current transaction.
+        /// </summary>
         public void Rollback()
         {
             string txUndoCatalogFile = Path.Combine(_core.Settings.TransactionPath, Id.ToString(), Constants.Filesystem.TransactionUndoCatalog);
@@ -132,17 +160,15 @@ namespace MamothDB.Server.Core.Models
 
             _latchKeys.TurnInAllKeys();
 
-            if (_session != null)
-            {
-                _session.CurrentTransaction = null;
-            }
-
             if (Directory.Exists(TransactionBackupPath))
             {
                 Directory.Delete(TransactionBackupPath, true);
             }
         }
 
+        /// <summary>
+        /// Writes the transaction catalog to disk. No caching.
+        /// </summary>
         public void CheckpointCatalog()
         {
             if (Directory.Exists(TransactionBackupPath) == false)
@@ -152,6 +178,10 @@ namespace MamothDB.Server.Core.Models
             _core.IO.PutJsonDirty(TransactionUndoCatalogFile, _undoActions);
         }
 
+        /// <summary>
+        /// Records the action of creating a directory to that it can be undone.
+        /// </summary>
+        /// <param name="filePath"></param>
         public void RecordCreateDirectory(string filePath)
         {
             var key = filePath.ToLower();
@@ -172,7 +202,7 @@ namespace MamothDB.Server.Core.Models
         }
 
         /// <summary>
-        /// Records the deletion of a directory. This method actually moves the folder to the undo location because that is the most efficient way to do it.
+        /// Records the deletion of a directory so that it can be undone. This method actually moves the folder to the undo location because that is the most efficient way to do it.
         /// </summary>
         /// <param name="filePath"></param>
         public void RecordDeleteDirectory(string filePath)
@@ -198,6 +228,10 @@ namespace MamothDB.Server.Core.Models
             }
         }
 
+        /// <summary>
+        /// Records the action of writing to a file so that it can be undone.
+        /// </summary>
+        /// <param name="filePath"></param>
         public void RecordFileWrite(string filePath)
         {
             var key = filePath.ToLower();
